# cmd/api/openapi.yaml
openapi: 3.1.0
info:
  title: Recurso API
  version: 1.0.0
  description: |
    The Recurso billing engine REST API.

    Authenticate by passing your API key as a bearer token:

        Authorization: Bearer <api_key>

    Obtain an API key by registering a tenant via `POST /auth/register`.
    All authenticated endpoints live under the `/v1` prefix. Mutating
    endpoints support idempotency via the `X-Idempotency-Key` header.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://billing.example.com
    description: Example deployment — substitute the base URL of your own Recurso deployment.
security:
  - bearerAuth: []
...
paths:
...
  /v1/users:
    get:
      operationId: listUsers
      tags: [Team]
      summary: List team members
      description: Lists all dashboard users in the caller's tenant. Any authenticated member may read.
      responses:
        '200':
          description: Users in the tenant.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createUser
      tags: [Team]
      summary: Add a team member
      description: Creates a teammate. Requires the caller to be owner or admin (or a tenant API key).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, role, password]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                role:
                  type: string
                  enum: [owner, admin, member]
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '201':
          description: Teammate created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/users/{id}:
    patch:
      operationId: updateUserRole
      tags: [Team]
      summary: Change a team member's role
      description: Updates a teammate's role. Owner/admin only. The last owner cannot be demoted.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [owner, admin, member]
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteUser
      tags: [Team]
      summary: Remove a team member
      description: Removes a teammate. Owner/admin only. Cannot remove the last owner or your own account.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: Tenant API key obtained from `POST /auth/register` or `POST /v1/developer/keys`.
    portalSession:
      type: apiKey
      in: cookie
      name: portal_session
      description: Customer-portal session cookie issued by `GET /portal/auth/verify`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: recurso_session
      description: >
        Dashboard user session cookie (httpOnly) issued by `POST /auth/register`
        and `POST /auth/login`. v1 endpoints accept EITHER this cookie or the
        tenant API key (`bearerAuth`).
...
  responses:
    BadRequest:
      description: The request body or parameters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials (API key or session cookie).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted (insufficient role).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The request conflicts with current state (e.g. duplicate email, last owner).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          description: Structured error detail.
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              examples: [BAD_REQUEST, UNAUTHORIZED, NOT_FOUND, CONFLICT, VALIDATION_ERROR, INVALID_API_KEY, RATE_LIMITED, INTERNAL_ERROR]
            message:
              type: string
              description: Human-readable explanation.
...
    User:
      type: object
      description: A dashboard user account within a tenant.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
      required: [id, email, name, role]